+++ resources/assets/js/components-vue/MediaGallery.vue
@@ -4,13 +4,18 @@
             <h3 class="h3 h1-dark">Kies een bestaande afbeelding</h3>
 
             <div class="flex items-center justify-end space-x-4">
-                <input placeholder="Zoek op bestandsnaam ..." type="text" v-model="searchQuery" />
+                <input
+                    v-model="searchQuery"
+                    type="text"
+                    placeholder="Zoek op bestandsnaam ..."
+                    class="form-input-field"
+                />
 
                 <button class="btn btn-primary" @click.prevent="search()">Filter</button>
             </div>
         </div>
 
-        <div data-overflow-scroll class="-mx-3 overflow-scroll row max-h-[50vh]">
+        <div data-overflow-scroll class="-mx-3 overflow-scroll row-start-start max-h-[50vh]">
             <div class="flex justify-center w-full" v-if="isLoading && assets.length < 1">
                 <svg
                     width="24"
@@ -158,6 +163,7 @@ export default {
                     .get(`${this.url}?limit=${this.limit}&excluded=${this.uploaded}&search=${this.searchQuery}`)
                     .then((response) => {
                         this.assets = response.data;
+                        console.log(this.assets);
                         this.isLoading = false;
                     })
                     .catch((errors) => {