+++ CHANGELOG.md
@@ -9,6 +9,7 @@ principles.
 -   Fixed: MediaGallery images were not showing up in the expected row layout
 -   Fixed: fragment titles in fragment-select-existing sidebar had h1 styling
 -   Fixed: bug where model has other primary than default id column
+-   Fixed: detaching asset with entity id being uuid failed (upgraded assetlibrary to 0.8.6)
 
 ## 0.7.20 - 2023-02-21